add main loop and changed all floats to double now the planes are moving on single machine

+++ ParallelProject/PointXY.h
@@ -3,9 +3,9 @@ class PointXY
 {
 public:
 	PointXY(void);
-	PointXY(int x,int y);
+	PointXY(double x,double y);
 	~PointXY(void);
-	int x;
-	int y;
+	double x;
+	double y;
 };
